requirements.txt
flask~=3.1.0
yfinance~=0.2.58
matplotlib
seaborn
transformers
tensorflow~=2.19.0
flask_cors~=5.0.1
numpy~=2.1.3
plotly~=6.0.1
keras
pandas~=2.2.3
scikit-learn~=1.6.1
textblob~=0.19.0
requests~=2.32.3
django-googlesearch~=0.2
beautifulsoup4~=4.13.4
groq~=0.24.0